refacto location tooltip
